+++ src/components/layouts/admin-layout/top-menu-bar/NotificationSymbol.tsx
@@ -1,0 +1,15 @@
+import IconButton from "@mui/material/IconButton";
+import Badge from "@mui/material/Badge";
+import NotificationsIcon from "@mui/icons-material/Notifications";
+
+export const NotificationSymbol = () => {
+  return (
+    <>
+      <IconButton color="inherit">
+        <Badge badgeContent={4} color="secondary">
+          <NotificationsIcon />
+        </Badge>
+      </IconButton>
+    </>
+  );
+};
