+++ src/components/reports/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Reports: React.FC = () => {
+  return <h1>Reports</h1>;
+};
+
+export default Reports;
